requirements.txt
Django~=2.0.13
django-friendship==1.8.2
django-crispy-forms==1.9.0
pillow==7.1.1
